+++ src/app/admin/page.tsx
@@ -47,7 +47,7 @@ export default function Admin() {
         </Link>
       </SousTitre1>
       <SousTitre1>
-        <Link className="underline" href="/admin/submit-screenings">
+        <Link className="underline" href="/admin/rajouter-des-seances">
           Rajouter des séances
         </Link>
       </SousTitre1>
